+++ pairtools_parquet/lib/__init__.py
@@ -11,6 +11,7 @@ from . import (
     filterbycov,
     flip,
     markasdup,
+    phase,
     restrict,
     sample,
     merge,